requirements.txt
aws-cdk-lib>=2.242.0,<2.1100.0
constructs>=10.0.0,<11.0.0
boto3>=1.34.0
aws-cdk.aws-apigatewayv2-alpha>=2.114.1a0
aws-cdk.aws-apigatewayv2-integrations-alpha>=2.114.1a0
cdk-nag>=2.28.0
aws-cdk.lambda-layer-kubectl-v31>=2.0.0
# aws-cdk.aws-bedrock-agentcore-alpha — uncomment when the CDK construct stabilizes
# aws-cdk.aws-bedrock-agentcore-alpha>=2.242.0a0
